pyproject.toml
[tool.poetry]
name = "src"
version = "0.1.0"
description = ""
authors = ["Yangxinyu <xinyux@wharton.upenn.edu>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.11.6"
openai = "^1.3.7"
torch = "^2.1.1"
faiss-cpu = "^1.7.4"
sentence-transformers = "^2.2.2"
numpy = "^1.26.2"
geopandas = "^0.14.1"
python-dotenv = "^1.0.0"
streamlit = "^1.29.0"
watchdog = "^3.0.0"
beautifulsoup4 = "^4.12.3"
geopy = "^2.4.1"
clipboard = "^0.0.4"
plotly = "^5.18.0"
matplotlib = "^3.8.2"
rouge = "^1.0.1"
kaleido = "0.2.1"
pydeck = "^0.9.1"
shapely = "^2.0.4"
pygris = "^0.1.6"
census = "^0.8.22"
us = "^3.1.1"
setuptools = "^70.0.0"
statsmodels = "^0.14.2"
bert-score = "^0.3.13"
wordcloud = "^1.9.3"
folium = "^0.19.4"
streamlit-folium = "^0.24.0"
bcrypt = "^5.0.0"
fpdf2 = "^2.8.5"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
